backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.3.4
python-dotenv==1.0.1
python-multipart==0.0.20
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.36
pgvector==0.3.2
numpy==1.26.4
scikit-learn==1.5.2
hdbscan==0.8.38.post1
google-generativeai==0.8.3
